refactor: Remove last duplicate 'server'
