+++ src/influxdb_ioxd/rpc/service.rs
@@ -23,9 +23,10 @@ use generated_types::{node, Node};
 use query::exec::fieldlist::FieldList;
 use query::group_by::GroupByAndAggregate;
 
-use crate::server::org_and_bucket_to_database;
-use crate::server::rpc::expr::{self, AddRPCNode, Loggable, SpecialTagKeys};
-use crate::server::rpc::input::GrpcInputs;
+use super::super::names::org_and_bucket_to_database;
+use super::expr::{self, AddRPCNode, Loggable, SpecialTagKeys};
+use super::input::GrpcInputs;
+
 use data_types::DatabaseName;
 
 use query::{
@@ -110,13 +111,13 @@ pub enum Error {
     #[snafu(display("Error converting Predicate '{}: {}", rpc_predicate_string, source))]
     ConvertingPredicate {
         rpc_predicate_string: String,
-        source: crate::server::rpc::expr::Error,
+        source: super::expr::Error,
     },
 
     #[snafu(display("Error converting group type '{}':  {}", aggregate_string, source))]
     ConvertingReadGroupType {
         aggregate_string: String,
-        source: crate::server::rpc::expr::Error,
+        source: super::expr::Error,
     },
 
     #[snafu(display(
@@ -126,7 +127,7 @@ pub enum Error {
     ))]
     ConvertingReadGroupAggregate {
         aggregate_string: String,
-        source: crate::server::rpc::expr::Error,
+        source: super::expr::Error,
     },
 
     #[snafu(display(
@@ -136,7 +137,7 @@ pub enum Error {
     ))]
     ConvertingWindowAggregate {
         aggregate_string: String,
-        source: crate::server::rpc::expr::Error,
+        source: super::expr::Error,
     },
 
     #[snafu(display("Error computing series: {}", source))]
@@ -149,14 +150,10 @@ pub enum Error {
     ComputingGroupedSeriesSet { source: SeriesSetError },
 
     #[snafu(display("Error converting time series into gRPC response:  {}", source))]
-    ConvertingSeriesSet {
-        source: crate::server::rpc::data::Error,
-    },
+    ConvertingSeriesSet { source: super::data::Error },
 
     #[snafu(display("Converting field information series into gRPC response:  {}", source))]
-    ConvertingFieldList {
-        source: crate::server::rpc::data::Error,
-    },
+    ConvertingFieldList { source: super::data::Error },
 
     #[snafu(display("Error sending results via channel:  {}", source))]
     SendingResults {
@@ -1159,7 +1156,7 @@ where
 
 #[cfg(test)]
 mod tests {
-    use crate::server::rpc::id::ID;
+    use super::super::id::ID;
 
     use super::*;
     use arrow_deps::arrow::datatypes::DataType;
@@ -1729,7 +1726,7 @@ mod tests {
         // Note we don't include the actual line / column in the
         // expected panic message to avoid needing to update the test
         // whenever the source code file changed.
-        let expected_error = "panicked at 'This is a test panic', src/server/rpc/service.rs";
+        let expected_error = "panicked at 'This is a test panic', src/influxdb_ioxd/rpc/service.rs";
         assert!(
             captured_logs.contains(expected_error),
             "Logs did not contain expected panic message '{}'. They were\n{}",
